Transliterated for TypeScript
Set the code for TypeScript.

Removed the concept of plugins, I suggest the use of Decorators, that is, those functions that can be transacted or encrypted, should have a note of the corresponding type, with appropriate parameterization.

Small adjustment in the package.json file by entering my contribution in this fork and also correcting the use of the license in the package.